feat: 공지사항관련 기능 테스트

+++ src/main/java/org/myteam/server/notice/service/NoticeCommentService.java
@@ -43,6 +43,7 @@ public class NoticeCommentService {
      * 공지사항 댓글 생성
      */
     public NoticeCommentSaveResponse save(Long noticeId, NoticeCommentSaveRequest request, String createdIp) {
+        log.info("공지사항: {}의 댓글 생성 시도", noticeId);
         Notice notice = noticeReadService.findById(noticeId);
         Member member = securityReadService.getMember();
 
@@ -54,6 +55,8 @@ public class NoticeCommentService {
 
         boolean isRecommended = noticeRecommendReadService.isRecommended(noticeComment.getId(), member.getPublicId());
 
+        log.info("공지사항: {}의 댓글 생성 성공", noticeId);
+
         return NoticeCommentSaveResponse.createResponse(noticeComment, member, isRecommended);
     }
 
@@ -61,17 +64,20 @@ public class NoticeCommentService {
      * 공지사항 댓글 수정
      */
     public NoticeCommentSaveResponse update(Long noticeCommentId, NoticeCommentUpdateRequest request) {
+        log.info("공지사항 댓글: {} 수정 시도", noticeCommentId);
         Member member = securityReadService.getMember();
         NoticeComment noticeComment = noticeCommentReadService.findById(noticeCommentId);
 
         noticeComment.verifyNoticeCommentAuthor(member);
-        if (!MediaUtils.verifyImageUrlAndRequestImageUrl(noticeComment.getImageUrl(), request.getImageUrl())) {
+        if (MediaUtils.verifyImageUrlAndRequestImageUrl(noticeComment.getImageUrl(), request.getImageUrl())) {
             s3Service.deleteFile(MediaUtils.getImagePath(request.getImageUrl()));
         }
 
         noticeComment.updateComment(request.getImageUrl(), badWordFilter.filterMessage(request.getComment()));
 
         boolean isRecommended = noticeCommentRecommendReadService.isRecommended(noticeComment.getId(), member.getPublicId());
+
+        log.info("공지사항 댓글: {} 수정 성공", noticeCommentId);
 
         return NoticeCommentSaveResponse.createResponse(noticeComment, member, isRecommended);
     }
@@ -80,13 +86,16 @@ public class NoticeCommentService {
      * 공지사항 댓글 삭제
      */
     public void deleteNoticeComment(Long noticeCommentId) {
+        log.info("공지사항 댓글: {} 삭제 시도", noticeCommentId);
         Member member = securityReadService.getMember();
         NoticeComment noticeComment = noticeCommentReadService.findById(noticeCommentId);
 
         noticeComment.verifyNoticeCommentAuthor(member);
 
         // S3 이미지 삭제
-        s3Service.deleteFile(MediaUtils.getImagePath(noticeComment.getImageUrl()));
+        if (noticeComment.getImageUrl() != null) {
+            s3Service.deleteFile(MediaUtils.getImagePath(noticeComment.getImageUrl()));
+        }
         // 대댓글 삭제 (카운트도 포함)
         int minusCount = deleteNoticeReply(noticeComment.getId());
         // 댓글 삭제
@@ -94,6 +103,8 @@ public class NoticeCommentService {
 
         // 댓글 카운트 감소
         noticeCountService.minusCommentCount(noticeComment.getNotice().getId(), minusCount + 1);
+
+        log.info("공지사항 댓글: {} 삭제 성공", noticeCommentId);
     }
 
     /**
